Add modules for di
