+++ app/src/main/java/com/example/kotlinexample/KotlinSampleApplication.kt
@@ -1,0 +1,7 @@
+package com.example.kotlinexample
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class KotlinSampleApplication : Application()
