+++ app/src/main/java/com/example/kotlinexample/di/DatabaseModule.kt
@@ -1,0 +1,23 @@
+package com.example.kotlinexample.di
+
+import android.content.Context
+import com.example.kotlinexample.room.SampleDatabase
+import com.example.kotlinexample.search.RepositoryDao
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ApplicationComponent
+import dagger.hilt.android.qualifiers.ApplicationContext
+
+@Module
+@InstallIn(ApplicationComponent::class)
+object DatabaseModule {
+
+    @Provides
+    fun getDatabase(@ApplicationContext context: Context): SampleDatabase =
+        SampleDatabase.getInstance(context)
+
+    @Provides
+    fun getRepositoryDao(@ApplicationContext context: Context): RepositoryDao =
+        SampleDatabase.getInstance(context).repositoryDao()
+}
