+++ app/src/main/java/com/example/kotlinexample/network/SampleRetrofit.kt
@@ -13,12 +13,12 @@ object SampleRetrofit {
 
     fun <T : Any> create(
         service: KClass<T>,
-        client: OkHttpClient = OkHttpProvider.gitHubApi
+        client: OkHttpClient
     ): T = create(service.java, client)
 
     fun <T> create(
         service: Class<T>,
-        client: OkHttpClient = OkHttpProvider.gitHubApi
+        client: OkHttpClient
     ): T = Retrofit.Builder()
         .baseUrl(BASE_URL)
         .client(client)
