+++ app/src/main/java/com/example/kotlinexample/room/SampleDatabase.kt
@@ -45,5 +45,5 @@ abstract class SampleDatabase : RoomDatabase() {
         }
     }
 
-    abstract fun searchDao(): RepositoryDao
+    abstract fun repositoryDao(): RepositoryDao
 }
